fix task solution
